Refactor: extract index package
